AB#5819 - Moved composing of core/infrastructure items into CoreInitialComposer
