+++ src/Umbraco.Core/Actions/ActionCollectionBuilder.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using Umbraco.Core.Composing;
 namespace Umbraco.Web.Actions
 {
-    internal class ActionCollectionBuilder : LazyCollectionBuilderBase<ActionCollectionBuilder, ActionCollection, IAction>
+    public class ActionCollectionBuilder : LazyCollectionBuilderBase<ActionCollectionBuilder, ActionCollection, IAction>
     {
         protected override ActionCollectionBuilder This => this;
